move zerotier-client to pkg
