+++ pkg/zerotier-client/ztutil/identity.go
