+++ pkg/zerotier/resource_identity.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	ztutil "github.com/someara/terraform-provider-zerotier/zerotier-client/ztutil"
+	ztutil "github.com/someara/terraform-provider-zerotier/pkg/zerotier-client/ztutil"
 )
 
 func resourceIdentity() *schema.Resource {
